requirements.txt
langchain
langchain-voyageai
python-dotenv
google-cloud-storage